BE-586 | Claimbot (#524)
* BE-586 | Claimbot prototype

Init

* BE-586 | WIP

* BE-586 | WIP

* BE-595 | Clean up

* BE-586 | Add docs

* BE-586 | Clean up

* BE-586 | Add docs

* BE-596 | Tests init

* BE-586 | Add tests for slices, orderbook packages

* BE-586 | claimbot/tx tests

* BE-586 | claimbot/order.go tests

* BE-586 | Requested changes

* BE-586 | Process block orderbooks, tests

* BE-586 | Requested changes

* BE-586 | Config update

* BE-586 | OrderBookClient use slices.Split for pagination

Cleans up OrderBookClient by reusing slices.Split instead of duplicating
splitting slices into chunks logic in some of the methods.

* BE-586 | Clean up

* BE-586 | Fix fillbot docker-compose

Fixes errors running fillbot via docker-compose

* BE-586 | Docs, docker compose fixes

* BE-586 | Run fillbot via docker-compose

* BE-586 | Run claimbot via docker-compose, clean up

* BE-586 | Cleanup

* BE-586 | Named logger

* BE-586 | Requested changes

* BE-586 | Logging failing tx

* BE-586 | Increase gas adjustment

* BE-586 | Error logging fix

* BE-586 | Trace name update

* BE-586 | Requested changes #1

* BE-586 | Requested changes #2

* BE-586 | Sequence number update

* BE-586 | added tests

* BE-586 | Suggested improvements

+++ ingest/usecase/plugins/orderbook/fillbot/create_copy_config.sh
@@ -1,8 +1,18 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 # Define the input and output file paths
 ORIGINAL_APP_TOML_NAME="$HOME/.osmosisd/config/app.toml"   # Replace with the actual file path
 BACKUP_APP_TOML_NAME="$HOME/.osmosisd/config/app-backup.toml"
+
+if [ ! -f "$ORIGINAL_APP_TOML_NAME" ]; then
+	echo "Error: Source file $ORIGINAL_APP_TOML_NAME does not exist."
+	exit 1
+fi
+	
+if [ -f $BACKUP_APP_TOML_NAME ]; then
+	echo "Backup file $BACKUP_APP_TOML_NAME already exist, no modifications will be made."
+	exit 0
+fi
 
 mv $ORIGINAL_APP_TOML_NAME $BACKUP_APP_TOML_NAME
